+++ source/ProcessManager.Example.Orchestrations.Abstrations/Processes/BRS_X03_ActorRequestProcessExample/V1/Model/ActorRequestProcessExampleEnqueueRejectedDataV1.cs
@@ -12,16 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 using Energinet.DataHub.ProcessManager.Components.Abstractions.BusinessValidation;
 
-namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X03_ActorRequestProcessExample.V1;
+namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X03_ActorRequestProcessExample.V1.Model;
 
 /// <summary>
-/// Input to BRS-X03
+/// Data when enqueueing BRS-X03 actor messages.
 /// </summary>
-public record ActorRequestProcessExampleInputV1(
-    string RequestedByActorNumber,
-    string RequestedByActorRole,
-    string BusinessReason)
-        : IInputParameterDto, IBusinessValidatedDto;
+public record ActorRequestProcessExampleEnqueueRejectedDataV1(
+    List<ValidationErrorDto> ValidationErrors);
