+++ source/ProcessManager.Example.Orchestrations.Abstrations/Processes/BRS_X03_ActorRequestProcessExample/V1/Model/ActorRequestProcessExampleNotifyEventV1.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 
-namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X03_ActorRequestProcessExample.V1;
+namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X03_ActorRequestProcessExample.V1.Model;
 
 public record ActorRequestProcessExampleNotifyEventV1(
     string OrchestrationInstanceId)
